- A lot of reorganization in the UI based on the feature that the code is part of - Fixing initial strava conversions, so we set them in the unit that we like - Converting Imperial system displayed measurements back into metric (which is the DB default value) - Adding DELETE Activity
