+++ packages/api/src/exercises/types.ts
@@ -14,7 +14,7 @@ export class ExerciseActivityUpload {
 export interface IStubAttributeOptions {
 	type?: EnumAttributeType;
 	unit?: EnumUnitTypes;
-	unitDisplay?: EnumUnitTypes;
+	unitConversion?: EnumUnitTypes;
 }
 
 export interface IUploadStrava {
