+++ packages/ui/src/App.vue
@@ -4,9 +4,9 @@ import { useRoute } from "vue-router";
 import { isAuthenticated } from "@/authentication.ts";
 import LoadingMask from "@/components/LoadingMask.vue";
 import { useGlobalError } from "@/composables/app.ts";
-import { RouteExercises } from "@/router/exercises.ts";
-import { RouteJobs } from "@/router/jobs.ts";
+import { RouteExercises } from "@/views/exercises/constants.ts";
 import NavigationExercises from "@/views/exercises/NavigationExercises.vue";
+import { RouteJobs } from "@/views/jobs/constants.ts";
 import NavigationJobs from "@/views/jobs/NavigationJobs.vue";
 import NavigationProducts from "@/views/NavigationProducts.vue";
 import NavigationUser from "@/views/NavigationUser.vue";
