+++ packages/ui/src/router.ts
@@ -1,8 +1,9 @@
 ﻿import { createRouter, createWebHashHistory, type RouteRecordRaw } from "vue-router";
 import { isAuthenticated } from "@/authentication.ts";
-import { ExercisesRoutes } from "@/router/exercises.ts";
-import { JobRoutes, RouteJobApplications } from "@/router/jobs.ts";
-import { restoreRoute, RouteHome, RouteLogin } from "@/router/routes.ts";
+import { restoreRoute, RouteHome, RouteLogin } from "@/constants.ts";
+import { ExercisesRoutes } from "@/views/exercises/routes.ts";
+import { RouteJobApplications } from "@/views/jobs/constants.ts";
+import { JobRoutes } from "@/views/jobs/routes.ts";
 import ViewLogIn from "@/views/ViewLogIn.vue";
 
 export const routes: RouteRecordRaw[] = [{
