+++ packages/ui/src/views/exercises/ViewExercisesHistory.vue
@@ -1,5 +1,5 @@
 ﻿<script setup lang="ts">
-import { useListExercisesHistory } from "@/composables/audits.ts";
+import { useListExercisesHistory } from "@/views/exercises/composables/activities.ts";
 import ViewFeatureHistory from "@/views/ViewFeatureHistory.vue";
 
 const { data } = useListExercisesHistory();
