+++ packages/ui/src/views/exercises/composables/activities.ts
@@ -1,7 +1,18 @@
-﻿import { onUnmounted, ref, unref } from "vue";
-import { EnumActivitySource, type ExerciseActivityCreateViewModel } from "@incutonez/life-stats-spec";
+﻿import { ref, unref } from "vue";
+import {
+	EnumActivitySource,
+	type ExerciseActivityCreateViewModel,
+	type ExerciseActivityViewModel,
+} from "@incutonez/life-stats-spec";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/vue-query";
-import { ActivitiesAPI, QueryKeyActivities, QueryKeyExercises } from "@/views/exercises/constants.ts";
+import { ExercisesAPI } from "@/api.ts";
+import { getInvalidateQueryPredicate, useInvalidateQueries } from "@/composables/app.ts";
+import {
+	ActivitiesAPI,
+	QueryKeyActivities,
+	QueryKeyExercises,
+	QueryListExercisesAudits,
+} from "@/views/exercises/constants.ts";
 
 interface IImportMutation {
 	file: File;
@@ -16,6 +27,32 @@ export function useListActivities() {
 			return data.data;
 		},
 	});
+}
+
+export function useDeleteActivity() {
+	const deletingRecord = ref(false);
+	const queryClient = useQueryClient();
+	const deleteMutation = useMutation({
+		async mutationFn(record: ExerciseActivityViewModel) {
+			return ActivitiesAPI.deleteActivity(record.id);
+		},
+		async onSuccess() {
+			await queryClient.invalidateQueries(getInvalidateQueryPredicate(QueryKeyExercises));
+		},
+	});
+
+	async function deleteRecord(record?: ExerciseActivityViewModel) {
+		if (record) {
+			deletingRecord.value = true;
+			await deleteMutation.mutateAsync(record);
+			deletingRecord.value = false;
+		}
+	}
+
+	return {
+		deletingRecord,
+		deleteRecord,
+	};
 }
 
 export function useImportActivities() {
@@ -50,13 +87,15 @@ export function useImportActivities() {
 
 export function useUploadActivities() {
 	const addedRecords = ref<ExerciseActivityCreateViewModel[]>([]);
-	const queryClient = useQueryClient();
 	const addingRecords = ref(false);
 	const added = ref(false);
 	const updateMutation = useMutation({
 		async mutationFn(records: ExerciseActivityCreateViewModel[]) {
 			addingRecords.value = true;
-			await ActivitiesAPI.uploadActivities(records);
+			await ActivitiesAPI.uploadActivities(records, {
+				// Set a 2 minute timeout, just in case it's a very large upload
+				timeout: 120000,
+			});
 			addingRecords.value = false;
 		},
 	});
@@ -69,17 +108,21 @@ export function useUploadActivities() {
 		}
 	}
 
-	onUnmounted(() => {
-		if (added.value) {
-			queryClient.invalidateQueries({
-				queryKey: [QueryKeyExercises],
-			});
-		}
-	});
+	useInvalidateQueries(added, QueryKeyExercises);
 
 	return {
 		addedRecords,
 		addingRecords,
 		createApplications,
 	};
+}
+
+export function useListExercisesHistory() {
+	return useQuery({
+		queryKey: [QueryListExercisesAudits],
+		async queryFn() {
+			const { data } = await ExercisesAPI.getExercisesHistory();
+			return data.data;
+		},
+	});
 }
