+++ packages/ui/src/views/exercises/composables/routes.ts
@@ -1,0 +1,27 @@
+﻿import { useRouter } from "vue-router";
+import { useBaseRoutes } from "@/composables/routes.ts";
+import { RouteExercises, RouteExercisesActivities, RouteExercisesHistory } from "@/views/exercises/constants.ts";
+
+export function useExerciseRoutes() {
+	const router = useRouter();
+	const { isRouteSelected } = useBaseRoutes(router);
+
+	return {
+		isRouteSelected,
+		viewExercises() {
+			return router.push({
+				name: RouteExercises,
+			});
+		},
+		ViewActivities() {
+			return router.push({
+				name: RouteExercisesActivities,
+			});
+		},
+		viewHistory() {
+			return router.push({
+				name: RouteExercisesHistory,
+			});
+		},
+	};
+}
