+++ packages/ui/src/views/jobs/ViewApplication.vue
@@ -7,11 +7,11 @@ import FieldDate from "@/components/FieldDate.vue";
 import FieldText from "@/components/FieldText.vue";
 import { IconAdd, IconDelete, IconEdit, IconSave } from "@/components/Icons.ts";
 import TableData from "@/components/TableData.vue";
-import { provideApplicationRecord, useDeleteApplication } from "@/composables/applications.ts";
 import { useDateCreatedColumn, useDateUpdatedColumn, useTableActions, useTableData } from "@/composables/table.ts";
-import { useJobRoutes } from "@/router/jobs.ts";
 import { getUniqueId } from "@/utils/common.ts";
 import ViewCommentDialog from "@/views/jobs/applications/ViewCommentDialog.vue";
+import { provideApplicationRecord, useDeleteApplication } from "@/views/jobs/composables/applications.ts";
+import { useJobRoutes } from "@/views/jobs/composables/routes.ts";
 import FieldApplicationStatus from "@/views/jobs/shared/FieldApplicationStatus.vue";
 import FieldCompanies from "@/views/jobs/shared/FieldCompanies.vue";
 import DeleteDialog from "@/views/shared/DeleteDialog.vue";
