+++ packages/ui/src/views/jobs/constants.ts
@@ -1,0 +1,23 @@
+﻿export const QueryKeyJobs = "jobs-";
+
+export const QueryListJobAudits = `${QueryKeyJobs}listJobAudits`;
+
+export const QueryListApplications = `${QueryKeyJobs}listApplications`;
+
+export const QueryGetApplication = `${QueryKeyJobs}getApplication`;
+
+export const QueryListCompanies = `${QueryKeyJobs}listCompanies`;
+
+export const QueryGetCompanies = `${QueryKeyJobs}getCompanies`;
+
+export const RouteJobs = "jobs";
+
+export const RouteJobApplications = "jobs-applications";
+
+export const RouteJobApplication = "jobs-application";
+
+export const RouteJobCompanies = "jobs-companies";
+
+export const RouteJobCompanyApplication = "jobs-companies-application";
+
+export const RouteJobsHistory = "jobs-history";
