+++ packages/ui/src/views/jobs/shared/FieldCompanies.vue
@@ -1,10 +1,10 @@
 ﻿<script setup lang="ts">
 import type { CompanyViewModel } from "@incutonez/life-stats-spec";
 import FieldComboBoxVirtual from "@/components/FieldComboBoxVirtual.vue";
-import { useGetCompanies } from "@/composables/companies.ts";
 import { getCompanyRecords } from "@/stores/companies.ts";
 import { useAppSelector } from "@/stores/main.ts";
 import type { IExtendedComboBox } from "@/types/components.ts";
+import { useGetCompanies } from "@/views/jobs/composables/companies.ts";
 
 defineProps<IExtendedComboBox<CompanyViewModel>>();
 const companies = useAppSelector(getCompanyRecords);
